rename Shortcut, RichText, and Notification to have named file and index.js
